requirements.txt
Django==5.2.1
djangorestframework==3.16.0
djoser==2.3.1
djangorestframework-simplejwt==5.5.0
drf-nested-routers==0.94.2
tabulate==0.9.0
drf-yasg==1.21.10
factory_boy==3.3.3
flake8==7.3.0
black==25.1.0
isort==6.0.1